pyproject.toml
[project]
name = "idarve"
version = "0.1.0"
description = "Investigating Demographic Attribute Representation in Vision Encoders"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.5.0,<2.6",
    "torchvision>=0.20.0,<0.21",
    "torchmetrics>=1.8.0,<2.0",
    "lightning>=2.6.0,<3.0",
    "transformers>=5.1.0,<6.0",
    "datasets>=4.5.0,<5.0",
    "hydra-core>=1.3.2,<1.4",
    "omegaconf>=2.3.0,<2.4",
    "wandb>=0.24.0,<1.0",
    "h5py>=3.15.0,<4.0",
    "hdf5plugin>=6.0.0,<7.0",
    "timm>=1.0.0,<2.0",
    "scikit-learn>=1.8.0,<2.0",
    "umap-learn>=0.5.0,<1.0",
    "matplotlib>=3.10.0,<4.0",
    "seaborn>=0.13.0,<1.0",
    "pandas>=3.0.0,<4.0",
    "numpy>=2.4.0,<3.0",
    "scipy>=1.17.0,<2.0",
    "tqdm>=4.67.0,<5.0",
    "einops>=0.8.0,<1.0",
    "pillow>=10.2.0,<11.0",
    "bert-score>=0.3.13,<0.4",
    "concept-erasure>=0.2.4,<0.3",
    "statsmodels>=0.14.0,<0.15",
]

[project.optional-dependencies]
dev = [
    "jupyter",
    "ipykernel",
    "pytest>=8.0",
    "ruff>=0.9.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["src*"]

# ---------------------------------------------------------------------------
# Ruff (linting + formatting)
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "SIM108", # ternary operator
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"
